refactor api folder code organization and config.
