+++ api/http/access_token.go
@@ -1,0 +1,52 @@
+package http
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/konstellation-io/kli/internal/config"
+)
+
+// accessTokenResponse represents a response from the sign-in API.
+type accessTokenResponse struct {
+	Token string `json:"access_token"`
+}
+
+func GetAccessToken(cfg *config.Config, server *config.ServerConfig) (string, error) {
+	client := &http.Client{}
+	url := fmt.Sprintf("%s/api/v1/auth/token/signin", server.URL)
+
+	postData := bytes.NewBuffer([]byte(fmt.Sprintf(`{"apiToken":"%s"}`, server.APIToken)))
+
+	ctx, cancel := context.WithTimeout(context.Background(), cfg.DefaultRequestTimeout)
+	defer cancel()
+
+	req, err := http.NewRequestWithContext(ctx, "POST", url, postData)
+	if err != nil {
+		return "", err
+	}
+
+	req.Header.Set("Content-Type", "application/json")
+
+	r, err := client.Do(req)
+	if err != nil {
+		return "", err
+	}
+	defer r.Body.Close()
+
+	if r.Body == nil {
+		return "", ErrResponseEmpty
+	}
+
+	var t accessTokenResponse
+
+	err = json.NewDecoder(r.Body).Decode(&t)
+	if err != nil {
+		return "", err
+	}
+
+	return t.Token, nil
+}
